Created skeleton for ControlPanel component and subcomponents
