+++ src/adapters/driveAPI.js
@@ -2,6 +2,10 @@ import API, { BASE_URL } from './API';
 
 const DRIVE_URL = BASE_URL + '/drive'
 
+const getDrive = () => {
+  
+}
+
 const uploadFile = (file, progressCallback, completionCallback) => {
   const xhr = new XMLHttpRequest();
   xhr.open('POST', DRIVE_URL, true);
