+++ src/components/profile/ProfileWidget.js
@@ -1,0 +1,7 @@
+import React from 'react';
+
+const ProfileWidget = props => {
+  return(<div className="profile-widget" > </div>)
+}
+
+export default ProfileWidget;
